[HOTFIX] Fix how params dict is passed to request()
The params dict was being passed as a kwargs parameter (`**params`)
rather than as a kwarg itself (e.g., `params=params`).  As a result,
the change in the `request()` method was passing each of the parameters
intended for the YARN RM as keyword arguments to the `requests()` method
itself.  This produced the following error on such requests...

```
<class 'TypeError'> - 'request() got an unexpected keyword argument
'startedTimeBegin'
```

resulting in failures in `0.3.4`.