Honor http policy configured in yarn-site xml (#41)
In order to allow clients to have a means of using https for their
resource manager, this change enables that behavior ONLY in cases
where no resource manager address is provided.  In this case, the
module will first check if the http policy is HTTPS_ONLY and, in
such case, construct the appropriate endpoint.

Note that the changes are minimally evasive so as to prevent clients
from having to change code.